refactor(di): Create DataModule
