+++ core/src/main/kotlin/internal/di/DataModule.kt
@@ -17,10 +17,8 @@
  * along with ECInventory. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package ru.endlesscode.inventory.internal
+package ru.endlesscode.inventory.internal.di
 
-import kotlinx.serialization.hocon.Hocon
-import org.bukkit.plugin.Plugin
 import ru.endlesscode.inventory.InventoryLayout
 import ru.endlesscode.inventory.internal.config.ConfigurationHolder
 import ru.endlesscode.inventory.internal.data.DataHolder
@@ -30,33 +28,19 @@ import ru.endlesscode.inventory.internal.data.sql.SqliteDatabaseConfig
 import ru.endlesscode.inventory.internal.locale.I18N
 import ru.endlesscode.inventory.internal.locale.I18NBukkit
 import ru.endlesscode.inventory.slot.Slot
-import ru.endlesscode.mimic.bukkit.load
-import ru.endlesscode.mimic.items.BukkitItemsRegistry
+import java.nio.file.Path
 
-internal object DI {
-
-    private lateinit var plugin: Plugin
-    private val servicesManager get() = plugin.server.servicesManager
-    private val dataPath get() = plugin.dataFolder.toPath()
-
-    val scheduler: TaskScheduler by lazy { PluginTaskScheduler(plugin) }
-
-    private val itemsRegistry: BukkitItemsRegistry by lazy { checkNotNull(servicesManager.load()) }
-
-    val hocon: Hocon by lazy { Hocon { useConfigNamingConvention = true } }
+internal class DataModule(private val dataPath: Path) {
     val database: Database by lazy { Database(SqliteDatabaseConfig(dataPath)) }
 
-    private val configHolder by lazy { ConfigurationHolder(dataPath, MainConfigurationImpl.SERIALIZER) }
-    private val dataHolder by lazy { DataHolder(itemsRegistry, dataPath) }
-
     val locale: I18N by lazy { I18NBukkit(dataPath, config.locale) }
+
+    val configHolder by lazy { ConfigurationHolder(dataPath, MainConfigurationImpl.SERIALIZER) }
     val config: MainConfigurationImpl get() = configHolder.config
+
+    val dataHolder by lazy { DataHolder(dataPath) }
     val slots: Map<String, Slot> get() = dataHolder.slots
     val inventories: Map<String, InventoryLayout> get() = dataHolder.inventories
-
-    fun init(plugin: Plugin) {
-        this.plugin = plugin
-    }
 
     fun reload() {
         configHolder.reload()
