+++ core/src/main/kotlin/internal/listener/InventoryClicksRouter.kt
@@ -25,8 +25,8 @@ import org.bukkit.event.Listener
 import org.bukkit.event.inventory.*
 import org.bukkit.event.inventory.InventoryAction.*
 import ru.endlesscode.inventory.CustomInventory
-import ru.endlesscode.inventory.internal.DI
 import ru.endlesscode.inventory.internal.TaskScheduler
+import ru.endlesscode.inventory.internal.di.DI
 import ru.endlesscode.inventory.slot.InventorySlot
 
 /** Converts [InventoryInteractEvent] to [SlotInteraction] and passes it to inventory. */
@@ -82,7 +82,7 @@ internal class InventoryClicksRouter(
         // Issue: https://hub.spigotmc.org/jira/browse/SPIGOT-6145
         if (event.click == ClickType.SWAP_OFFHAND) {
             val playerInventory = event.whoClicked.inventory
-            scheduler.runTask { playerInventory.setItemInOffHand(playerInventory.itemInOffHand) }
+            scheduler.runOnMain { playerInventory.setItemInOffHand(playerInventory.itemInOffHand) }
         }
     }
